requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
python-telegram-bot==13.3
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
openpyxl == 3.0.6